정밀도가 중요한 금융 계산에서는 BigDecimal

+++ sharedPrompts/src/main/java/org/example/sharedprompts/domain/payment/config/RewardProperties.java
@@ -4,6 +4,8 @@ import lombok.Getter;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 /**
  * 리워드 설정 Properties (Immutable)
  * - 캐시백 및 포인트 적립률 관리
@@ -12,12 +14,12 @@ import org.springframework.stereotype.Component;
 @Component
 public class RewardProperties {
 
-    private final double cashbackRate;
-    private final double pointRate;
+    private final BigDecimal cashbackRate;
+    private final BigDecimal pointRate;
 
     public RewardProperties(
-            @Value("${payment.cashback.rate:0.01}") double cashbackRate,
-            @Value("${payment.point.rate:0.005}") double pointRate) {
+            @Value("${payment.cashback.rate:0.01}") BigDecimal cashbackRate,
+            @Value("${payment.point.rate:0.005}") BigDecimal pointRate) {
         this.cashbackRate = cashbackRate;
         this.pointRate = pointRate;
     }
